move: cookie file in bruno-request package
