+++ packages/bruno-common/src/index.ts
@@ -1,6 +1,5 @@
 export { mockDataFunctions } from './utils/faker-functions';
 export { default as interpolate } from './interpolate';
 export { default as isRequestTagsIncluded } from './tags';
-export { default as cookies } from './cookies';
 
 export * as utils from './utils';
