+++ packages/bruno-common/src/utils/index.ts
@@ -3,7 +3,3 @@ export {
   parseQueryParams,
   buildQueryString,
 } from './url';
-
-export {
-  isPotentiallyTrustworthyOrigin
-} from './url/validation';
