+++ packages/bruno-electron/src/store/cookies.js
@@ -1,5 +1,5 @@
 const Store = require('electron-store');
-const { cookies: cookiesModule } = require('@usebruno/common');
+const { cookies: cookiesModule } = require('@usebruno/requests');
 const { cookieJar } = cookiesModule;
 const { Cookie } = require('tough-cookie');
 const { createCookieString } = cookiesModule;
