+++ packages/bruno-requests/package.json
@@ -24,8 +24,9 @@
     "@grpc/grpc-js": "^1.13.3",
     "@grpc/proto-loader": "^0.7.15",
     "@types/qs": "^6.9.18",
+    "axios": "^1.9.0",
     "grpc-reflection-js": "^0.3.0",
-    "axios": "^1.9.0"
+    "is-ip": "^5.0.1"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.22.0",
@@ -37,8 +38,8 @@
     "@rollup/plugin-typescript": "^9.0.2",
     "@types/jest": "^29.5.11",
     "babel-jest": "^29.7.0",
-    "jest": "^29.2.0",
     "builtin-modules": "^5.0.0",
+    "jest": "^29.2.0",
     "rollup": "3.29.5",
     "rollup-plugin-dts": "^5.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
