+++ packages/bruno-requests/src/index.ts
@@ -1,5 +1,6 @@
 export { addDigestInterceptor, getOAuth2Token } from  './auth';
 export { GrpcClient, generateGrpcSampleMessage } from './grpc';
+export { default as cookies } from './cookies';
 
 export * as network from './network';
