+++ packages/bruno-requests/src/tests/utils/validation.spec.ts
@@ -1,4 +1,4 @@
-import { isPotentiallyTrustworthyOrigin } from './validation';
+import { isPotentiallyTrustworthyOrigin } from '../../utils/validation';
 
 describe('isPotentiallyTrustworthyOrigin', () => {
   describe('secure schemes', () => {
